requirements.txt
functions-framework==3.*
Flask==2.3.3
sendgrid==6.11.0
beautifulsoup4==4.12.3
feedparser==6.0.11
markdownify==0.13.1
requests==2.32.3
google-generativeai==0.8.2
python-dotenv==1.0.0
protobuf>=3.19.5,<5.0.0